refactor: Move parallel test runner code to a separate file
