+++ frappe/commands/utils.py
@@ -603,7 +603,7 @@ def run_tests(context, app=None, module=None, doctype=None, test=(), profile=Fal
 @click.option('--with-coverage', is_flag=True, help="Build coverage file")
 @pass_context
 def run_parallel_tests(context, app, build_number, total_builds, with_coverage):
-	from frappe.test_runner import ParallelTestRunner
+	from frappe.parallel_test_runner import ParallelTestRunner
 	site = get_site(context)
 	ParallelTestRunner(app, site=site, build_number=build_number, total_builds=total_builds, with_coverage=with_coverage)
